+++ concordium-std/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Make Write implementation for ContractStateTest resize the state automatically
   to be consistent with the Write implementation for ContractState.
+- Use little-endian encoding for sender contract addresses in receive contexts. This
+  reverts the change in concordium-std 0.4.1.
 
 ## concordium-std 0.4.1 (2021-02-22)